Move `ConvertQueryButton` component
