+++ frontend/src/metabase/query_builder/components/view/ViewHeader/ViewHeader.jsx
@@ -19,9 +19,9 @@ import { MODAL_TYPES } from "metabase/query_builder/constants";
 import { getDashboard } from "metabase/query_builder/selectors";
 import QuestionActions from "../../QuestionActions";
 import QuestionDescription from "../QuestionDescription";
-import ConvertQueryButton from "../ConvertQueryButton";
 import { FilterHeaderToggle, FilterHeader } from "../QuestionFilters";
 import { QuestionSummarizeWidget } from "../QuestionSummaries";
+import { ConvertQueryButton } from "./components/ConvertQueryButton";
 import { HeadBreadcrumbs } from "./components/HeaderBreadcrumbs";
 import { FilterHeaderButton } from "./components/FilterHeaderButton";
 import { ExploreResultsLink } from "./components/ExploreResultsLink";
