+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/ConvertQueryButton/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./ConvertQueryButton";
+export { ConvertQueryButton } from "./ConvertQueryButton";
